feat(tooltip): simplified tooltip component by making arrow a template
