+++ packages/tooltip/index.js
@@ -1,2 +1,1 @@
 export { LionTooltip } from './src/LionTooltip.js';
-export { LionTooltipArrow } from './src/LionTooltipArrow.js';
