--- packages/tooltip/lion-tooltip-arrow.js
@@ -1,3 +1,0 @@
-import { LionTooltipArrow } from './src/LionTooltipArrow.js';
-
-customElements.define('lion-tooltip-arrow', LionTooltipArrow);
